requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.10.0
django-environ==0.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-core==2.0.0
google-cloud-secret-manager==2.7.1
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
idna==3.2
libcst==0.3.21
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
